requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --config=pyproject.toml requirements/requirements.in
#
algorithm-x==0.1.0
    # via -r requirements/requirements.in
asgiref==3.7.2
    # via django
crispy-bootstrap5==0.7
    # via -r requirements/requirements.in
django==4.2.3
    # via
    #   -r requirements/requirements.in
    #   crispy-bootstrap5
    #   django-crispy-forms
    #   djangorestframework
django-crispy-forms==2.0
    # via
    #   -r requirements/requirements.in
    #   crispy-bootstrap5
djangorestframework==3.14.0
    # via -r requirements/requirements.in
pillow==10.0.0
    # via -r requirements/requirements.in
psycopg2-binary==2.9.6
    # via -r requirements/requirements.in
python-dotenv==1.0.0
    # via -r requirements/requirements.in
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.4
    # via django
